refactor(home): isolate dashboard data loading
