+++ client/src/components/HomePage.tsx
@@ -1,20 +1,15 @@
-import { tryGetGame } from "@shared/games/registry";
 import type { GameId, LapMeta, SessionMeta, SessionRecap as SessionRecapDto } from "@shared/types";
-import { useQueries } from "@tanstack/react-query";
 import { Link } from "@tanstack/react-router";
 import { Settings2 } from "lucide-react";
-import { type ReactNode, useEffect, useMemo, useState } from "react";
+import { type ReactNode } from "react";
 import { m } from "@/paraglide/messages";
 import type { DriverFingerprint } from "../../../server/ai/driver-profile-aggregate";
 import type { DriverProfileState } from "../hooks/queries";
-import { useDriverProfile, useDriverProfileRuns, useLaps, useSessionRecap, useSessions, useSettings, useTrackOutline, useTrackSectorBoundaries } from "../hooks/queries";
-import { client } from "../lib/rpc";
-import { getGameRoute, useGameId } from "../stores/game";
-import { useUiStore } from "../stores/ui";
 import { ActivityHeatmap } from "./ActivityHeatmap";
 import { formatLapTime } from "./LiveTelemetry";
-import { buildRecapText, SessionRecapView, type TrackOutlineData, type TrackSectorBounds } from "./SessionRecap";
+import { SessionRecapView, type TrackOutlineData, type TrackSectorBounds } from "./SessionRecap";
 import { Table, TBody, TD, TH, THead, TRow } from "./ui/AppTable";
+import { getGameRoute } from "../stores/game";
 import { balanceReading, controlLossReading, gripMedianReading, reversalsReading, type StyleTone } from "@shared/lib/style-readings";
 
 function StatCard({ label, value, sub, color }: { label: string; value: string; sub?: string; color?: string }) {
@@ -27,7 +22,7 @@ function StatCard({ label, value, sub, color }: { label: string; value: string; 
   );
 }
 
-function RecentLapsTable({ laps, carNames, trackNames, gameId }: { laps: LapMeta[]; carNames: Record<number, string>; trackNames: Record<number, string>; gameId: string | null }) {
+function RecentLapsTable({ laps, carNames, trackNames, gameId, onAnalyseLap }: { laps: LapMeta[]; carNames: Record<number, string>; trackNames: Record<number, string>; gameId: string | null; onAnalyseLap: (lap: LapMeta) => void }) {
   const showGame = !gameId; // show game column on global homepage
   if (laps.length === 0) {
     return <div className="p-6 text-center text-app-text/90-dim">{m.home_no_laps()}</div>;
@@ -52,8 +47,7 @@ function RecentLapsTable({ laps, carNames, trackNames, gameId }: { laps: LapMeta
             <TRow
               key={lap.id}
               onClick={() => {
-                if (!lap.gameId) return;
-                window.location.href = `${getGameRoute(lap.gameId)}/analyse?track=${lap.trackOrdinal ?? ""}&car=${lap.carOrdinal ?? ""}&lap=${lap.id}`;
+                onAnalyseLap(lap);
               }}
             >
               {showGame && (
@@ -96,6 +90,23 @@ function formatTimeAgo(date: Date): string {
   if (sec < 604800) return `${Math.floor(sec / 86400)}d ${m.home_days_ago()}`;
   return date.toLocaleDateString();
 }
+export type PeriodKey = "today" | "week" | "month" | "year" | "allTime";
+export type PeriodStats = Record<
+  PeriodKey,
+  {
+    laps: number;
+    valid: number;
+    best: number;
+    avgTime: number;
+    totalTime: number;
+    tracks: number;
+    cars: number;
+    sessions: number;
+    favCarOrd: number | null;
+    favCarCount: number;
+  }
+>;
+export type GameStats = Record<"fm" | "f1" | "acc" | "acEvo" | "iracing", { laps: number; time: string }>;
 
 export interface DriverProgressCardProps {
   gameId: string;
@@ -253,9 +264,6 @@ export function DriverProgressCard({
   );
 }
 
-type PeriodKey = "today" | "week" | "month" | "year" | "allTime";
-type PeriodStats = Record<PeriodKey, { laps: number; valid: number; best: number; avgTime: number; totalTime: number; tracks: number; cars: number; sessions: number; favCarOrd: number | null; favCarCount: number }>;
-type GameStats = Record<"fm" | "f1" | "acc" | "acEvo" | "iracing", { laps: number; time: string }>;
 export interface HomePageViewProps {
   gameId: GameId | null;
   gameDisplayName: string | null;
@@ -274,6 +282,11 @@ export interface HomePageViewProps {
   latestRecapBounds?: TrackSectorBounds;
   recapCopied: boolean;
   onCopyRecap: () => void;
+  onAnalyseLap: (lap: LapMeta) => void;
+  lapsLoading?: boolean;
+  lapsError?: boolean;
+  sessionsLoading?: boolean;
+  sessionsError?: boolean;
   onAnalyseRecap: () => void;
   periodTab: PeriodKey;
   periodStats: PeriodStats;
@@ -287,199 +300,6 @@ export interface HomePageViewProps {
   driverRunState?: DriverProfileState;
 }
 
-export function HomePage() {
-  const gameId = useGameId();
-  const gameAdapter = gameId ? tryGetGame(gameId) : null;
-  const { data: allLaps = [] } = useLaps();
-  const driverProfileQuery = useDriverProfile({ gameId });
-  const driverProfileRunsQuery = useDriverProfileRuns({ gameId });
-  const { data: sessions = [] } = useSessions();
-  const { displaySettings } = useSettings();
-  const { openSettings } = useUiStore();
-  const hiddenGames: string[] = displaySettings.hiddenGames ?? [];
-
-  const latestSession = useMemo(() => {
-    if (sessions.length === 0) return null;
-    return [...sessions].sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())[0];
-  }, [sessions]);
-  const { data: latestRecap, isLoading: latestRecapLoading, isError: latestRecapError } = useSessionRecap(latestSession?.id, latestSession?.gameId ?? null);
-  const { data: latestRecapOutline } = useTrackOutline(latestRecap?.trackOrdinal, latestRecap?.gameId ?? latestSession?.gameId ?? null);
-  const { data: latestRecapBounds } = useTrackSectorBoundaries(latestRecap?.trackOrdinal, latestRecap?.gameId ?? latestSession?.gameId ?? null);
-  const [recapCopied, setRecapCopied] = useState(false);
-
-  // Resolve car/track names for recent laps
-  const [carNames, setCarNames] = useState<Record<number, string>>({});
-  const [trackNames, setTrackNames] = useState<Record<number, string>>({});
-
-  const recentLaps = useMemo(
-    () =>
-      [...allLaps]
-        .filter((l) => l.lapTime > 0)
-        .sort((a, b) => new Date(b.createdAt).getTime() - new Date(a.createdAt).getTime())
-        .slice(0, 10),
-    [allLaps],
-  );
-  const medianLapSec = useMemo(() => {
-    const selectedLapTimes = driverProfileQuery.data?.selectedLapTimes ?? [];
-    const times = selectedLapTimes
-      .filter((lap) => lap.isValid && lap.lapTime > 0)
-      .map((lap) => lap.lapTime)
-      .sort((a, b) => a - b);
-    if (times.length === 0) return null;
-    const middle = Math.floor(times.length / 2);
-    return times.length % 2 === 0 ? (times[middle - 1] + times[middle]) / 2 : times[middle];
-  }, [driverProfileQuery.data]);
-
-  // Per-game stats — fetched from /api/stats per game so counts aren't
-  // capped by useLaps()'s 200-row limit (home and /<gameId> used to
-  // disagree when total laps across games exceeded 200).
-  const gameQueries = useQueries({
-    queries: (["fm-2023", "f1-2025", "acc", "ac-evo", "iracing"] as const).map((g) => ({
-      queryKey: ["stats", g],
-      queryFn: async () => {
-        const res = await client.api.stats.$get({ query: { gameId: g } });
-        if (!res.ok) throw new Error(res.statusText);
-        return res.json() as Promise<{ totalLaps: number; totalTimeSec: number }>;
-      },
-    })),
-  });
-
-  const gameStats = useMemo(() => {
-    const fmtTime = (sec: number) => {
-      if (sec <= 0) return "—";
-      const h = Math.floor(sec / 3600);
-      const m = Math.floor((sec % 3600) / 60);
-      return h > 0 ? `${h}h ${m}m` : `${m}m`;
-    };
-    const pick = (i: number) => {
-      const d = gameQueries[i].data;
-      return { laps: d?.totalLaps ?? 0, time: fmtTime(d?.totalTimeSec ?? 0) };
-    };
-    return { fm: pick(0), f1: pick(1), acc: pick(2), acEvo: pick(3), iracing: pick(4) };
-  }, [gameQueries]);
-
-  // Period metrics
-  const [periodTab, setPeriodTab] = useState<"today" | "week" | "month" | "year" | "allTime">("allTime");
-
-  const [{ todayStart, weekAgo, monthAgo, yearAgo }] = useState(() => {
-    const now = Date.now();
-    return {
-      todayStart: new Date().setHours(0, 0, 0, 0),
-      weekAgo: now - 7 * 24 * 60 * 60 * 1000,
-      monthAgo: now - 30 * 24 * 60 * 60 * 1000,
-      yearAgo: now - 365 * 24 * 60 * 60 * 1000,
-    };
-  });
-
-  const periodStats = useMemo(() => {
-    function computePeriod(laps: LapMeta[]) {
-      const valid = laps.filter((l) => l.isValid && l.lapTime > 0);
-      const best = valid.length > 0 ? Math.min(...valid.map((l) => l.lapTime)) : 0;
-      const avgTime = valid.length > 0 ? valid.reduce((s, l) => s + l.lapTime, 0) / valid.length : 0;
-      const totalTime = laps.reduce((s, l) => s + (l.lapTime > 0 ? l.lapTime : 0), 0);
-      const tracks = new Set(laps.map((l) => l.trackOrdinal).filter(Boolean)).size;
-      const cars = new Set(laps.map((l) => l.carOrdinal).filter(Boolean)).size;
-      const sessions = new Set(laps.map((l) => l.sessionId).filter(Boolean)).size;
-      const carCounts = new Map<number, number>();
-      for (const l of laps) {
-        if (l.carOrdinal) carCounts.set(l.carOrdinal, (carCounts.get(l.carOrdinal) ?? 0) + 1);
-      }
-      let favCarOrd: number | null = null;
-      let favCarCount = 0;
-      for (const [ord, count] of carCounts) {
-        if (count > favCarCount) {
-          favCarOrd = ord;
-          favCarCount = count;
-        }
-      }
-      return { laps: laps.length, valid: valid.length, best, avgTime, totalTime, tracks, cars, sessions, favCarOrd, favCarCount };
-    }
-
-    const gameLaps = gameId ? allLaps.filter((l) => l.gameId === gameId) : allLaps;
-
-    const todayLaps = gameLaps.filter((l) => new Date(l.createdAt).getTime() >= todayStart);
-    const weekLaps = gameLaps.filter((l) => new Date(l.createdAt).getTime() >= weekAgo);
-    const monthLaps = gameLaps.filter((l) => new Date(l.createdAt).getTime() >= monthAgo);
-    const yearLaps = gameLaps.filter((l) => new Date(l.createdAt).getTime() >= yearAgo);
-
-    return {
-      today: computePeriod(todayLaps),
-      week: computePeriod(weekLaps),
-      month: computePeriod(monthLaps),
-      year: computePeriod(yearLaps),
-      allTime: computePeriod(gameLaps),
-    };
-  }, [allLaps, gameId, todayStart, weekAgo, monthAgo, yearAgo]);
-
-  // Fetch names for recent laps + favourite cars
-  useEffect(() => {
-    const carOrds = [...new Set([...recentLaps.map((l) => l.carOrdinal), periodStats.today.favCarOrd, periodStats.week.favCarOrd, periodStats.month.favCarOrd].filter((o): o is number => o != null))];
-    const trackOrds = [...new Set(recentLaps.map((l) => l.trackOrdinal).filter((o): o is number => o != null))];
-    for (const ord of carOrds) {
-      if (carNames[ord]) continue;
-      // Find the gameId from a lap that has this ordinal
-      const lapForCar = recentLaps.find((l) => l.carOrdinal === ord);
-      client.api["car-name"][":ordinal"]
-        .$get({ param: { ordinal: String(ord) }, query: { gameId: (lapForCar?.gameId ?? gameId)! } })
-        .then((r) => (r.ok ? r.text() : ""))
-        .then((name) => setCarNames((prev) => ({ ...prev, [ord]: name })))
-        .catch(() => {});
-    }
-    for (const ord of trackOrds) {
-      if (trackNames[ord]) continue;
-      const lapForTrack = recentLaps.find((l) => l.trackOrdinal === ord);
-      client.api["track-name"][":ordinal"]
-        .$get({ param: { ordinal: String(ord) }, query: { gameId: (lapForTrack?.gameId ?? gameId)! } })
-        .then((r) => (r.ok ? r.text() : ""))
-        .then((name) => setTrackNames((prev) => ({ ...prev, [ord]: name })))
-        .catch(() => {});
-    }
-  }, [recentLaps, periodStats, gameId]);
-
-  const copyRecap = () => {
-    if (!latestRecap) return;
-    navigator.clipboard.writeText(buildRecapText(latestRecap)).then(() => {
-      setRecapCopied(true);
-      setTimeout(() => setRecapCopied(false), 1500);
-    });
-  };
-  const analyseRecap = () => {
-    if (!latestRecap || latestRecap.bestLapId == null) return;
-    window.location.href = `${getGameRoute(latestRecap.gameId)}/analyse?track=${latestRecap.trackOrdinal}&car=${latestRecap.carOrdinal}&lap=${latestRecap.bestLapId}`;
-  };
-  return (
-    <HomePageView
-      gameId={gameId}
-      gameDisplayName={gameAdapter?.displayName ?? null}
-      displaySettings={displaySettings}
-      allLaps={allLaps}
-      recentLaps={recentLaps}
-      carNames={carNames}
-      trackNames={trackNames}
-      gameStats={gameStats}
-      hiddenGames={hiddenGames}
-      latestSession={latestSession}
-      latestRecap={latestRecap}
-      latestRecapLoading={latestRecapLoading}
-      latestRecapError={latestRecapError}
-      latestRecapOutline={latestRecapOutline}
-      latestRecapBounds={latestRecapBounds}
-      recapCopied={recapCopied}
-      onCopyRecap={copyRecap}
-      onAnalyseRecap={analyseRecap}
-      periodTab={periodTab}
-      periodStats={periodStats}
-      onPeriodTabChange={setPeriodTab}
-      onOpenSettings={() => openSettings("games")}
-      driverGameId={gameId}
-      driverFingerprint={driverProfileQuery.data?.fingerprint ?? null}
-      driverLoading={driverProfileQuery.isLoading}
-      driverError={driverProfileQuery.error instanceof Error ? driverProfileQuery.error.message : driverProfileQuery.error ? String(driverProfileQuery.error) : null}
-      medianLapSec={medianLapSec}
-      driverRunState={driverProfileRunsQuery.data?.state}
-    />
-  );
-}
 
 export function HomePageView({
   gameId,
@@ -499,6 +319,7 @@ export function HomePageView({
   latestRecapBounds,
   recapCopied,
   onCopyRecap,
+  onAnalyseLap,
   onAnalyseRecap,
   periodTab,
   periodStats,
@@ -900,7 +721,7 @@ export function HomePageView({
 
             <section>
               <h2 className="mb-2 text-xs font-semibold uppercase tracking-wider text-app-text/90-muted">{m.home_recent_laps()}</h2>
-              <RecentLapsTable laps={recentLaps} carNames={carNames} trackNames={trackNames} gameId={gameId} />
+            <RecentLapsTable laps={recentLaps} carNames={carNames} trackNames={trackNames} gameId={gameId} onAnalyseLap={onAnalyseLap} />
             </section>
           </main>
 
@@ -985,7 +806,7 @@ export function HomePageView({
 
           <div>
             <h2 className="mb-2 text-xs font-semibold uppercase tracking-wider text-app-text/90-muted">{m.home_recent_laps()}</h2>
-            <RecentLapsTable laps={recentLaps} carNames={carNames} trackNames={trackNames} gameId={gameId} />
+            <RecentLapsTable laps={recentLaps} carNames={carNames} trackNames={trackNames} gameId={gameId} onAnalyseLap={onAnalyseLap} />
           </div>
         </>
       )}
